refactor(install): split finalize phase
